+++ board/sys_config/ti_msp_dl_config.c
@@ -498,11 +498,11 @@ SYSCONFIG_WEAK void SYSCFG_DL_BlueTooth_init(void)
     DL_UART_Main_init(BlueTooth_INST, (DL_UART_Main_Config *) &gBlueToothConfig);
     /*
      * Configure baud rate by setting oversampling and baud rate divisors.
-     *  Target baud rate: 115200
-     *  Actual baud rate: 115211.52
+     *  Target baud rate: 9600
+     *  Actual baud rate: 9600.24
      */
     DL_UART_Main_setOversampling(BlueTooth_INST, DL_UART_OVERSAMPLING_RATE_16X);
-    DL_UART_Main_setBaudRateDivisor(BlueTooth_INST, BlueTooth_IBRD_32_MHZ_115200_BAUD, BlueTooth_FBRD_32_MHZ_115200_BAUD);
+    DL_UART_Main_setBaudRateDivisor(BlueTooth_INST, BlueTooth_IBRD_32_MHZ_9600_BAUD, BlueTooth_FBRD_32_MHZ_9600_BAUD);