+++ board/sys_config/ti_msp_dl_config.h
@@ -160,9 +160,9 @@ extern "C" {
 #define GPIO_BlueTooth_IOMUX_TX                                  (IOMUX_PINCM12)
 #define GPIO_BlueTooth_IOMUX_RX_FUNC                   IOMUX_PINCM13_PF_UART0_RX
 #define GPIO_BlueTooth_IOMUX_TX_FUNC                   IOMUX_PINCM12_PF_UART0_TX
-#define BlueTooth_BAUD_RATE                                             (115200)
-#define BlueTooth_IBRD_32_MHZ_115200_BAUD                                   (17)
-#define BlueTooth_FBRD_32_MHZ_115200_BAUD                                   (23)
+#define BlueTooth_BAUD_RATE                                               (9600)
+#define BlueTooth_IBRD_32_MHZ_9600_BAUD                                    (208)
+#define BlueTooth_FBRD_32_MHZ_9600_BAUD                                     (21)
 /* Defines for Debug_Ex */
 #define Debug_Ex_INST                                                      UART1
 #define Debug_Ex_INST_FREQUENCY                                         32000000